+++ test/libsolidity/syntaxTests/types/contract_to_address.sol
@@ -1,0 +1,7 @@
+contract C {
+  function f() public view {
+    address a = address(this);
+    a;
+  }
+}
+// ----