+++ test/libsolidity/syntaxTests/types/contract_to_base_base.sol
@@ -1,0 +1,10 @@
+contract A {}
+contract B is A {}
+contract C is B {}
+contract D {
+  function f() public {
+    A a = new C();
+    a;
+  }
+}
+// ----