+++ test/libsolidity/syntaxTests/types/contract_to_unrelated_contract.sol
@@ -1,0 +1,9 @@
+contract A {}
+contract B {}
+contract C {
+  function f() public pure {
+    B b = B(new A());
+  }
+}
+// ----
+// TypeError: (80-90): Explicit type conversion not allowed from "contract A" to "contract B".